a.rename(b) should mutate a to look like it referred to b.
Unfortunately, the inmemory filesystem design disagrees with that.
Need to rethink this approach, maybe just stick to "a is not changed,
return something equal to b". But then we need tests to check that a
really is *not* changed.. and I think that behavior is much more
confusing. Then again, mutating paths is not nice either.